Moved admin code into package and set up configuration to import it.
