+++ api/admin/templates.py
@@ -1,14 +1,16 @@
+admin = """
 <!doctype html>
 <html>
 <head><title>Circulation Manager</title></head>
 <body>
-  <script src="/static/circulation-web.js"></script>
+  <script src=\"/admin/static/circulation-web.js\"></script>
   <script>
     var circulationWeb = new CirculationWeb(
       {% if csrf_token %}
-        "{{ csrf_token }}"
+        \"{{ csrf_token }}\"
       {% endif %}
     );
   </script>
 </body>
 </html>
+"""
