+++ api/controller.py
@@ -6,8 +6,6 @@ import time
 import urllib
 import urlparse
 import uuid
-import os
-import base64
 
 from lxml import etree
 
@@ -83,7 +81,6 @@ from config import (
 from lanes import make_lanes
 
 from adobe_vendor_id import AdobeVendorIDController
-from oauth import GoogleAuthService
 from axis import (
     Axis360API,
 )
@@ -100,6 +97,7 @@ from circulation import (
     CirculationAPI,
     DummyCirculationAPI,
 )
+
 
 class CirculationManager(object):
 
@@ -208,7 +206,6 @@ class CirculationManager(object):
         self.accounts = AccountController(self)
         self.urn_lookup = URNLookupController(self._db)
         self.work_controller = WorkController(self)
-        self.admin_controller = AdminController(self)
 
         self.heartbeat = HeartbeatController()
         self.service_status = ServiceStatusController(self)
@@ -397,99 +394,6 @@ class CirculationManagerController(object):
         return None
 
 
-class AdminController(CirculationManagerController):
-
-    ERROR_RESPONSE_TEMPLATE = """<!DOCTYPE HTML>
-<html lang="en">
-<head><meta charset="utf8"></head>
-</body>
-<p><strong>%(status_code)d ERROR:</strong> %(message)s</p>
-</body>
-</html>"""
-
-    @property
-    def google(self):
-        return GoogleAuthService.from_environment(
-            self.url_for('google_auth_callback'), test_mode=self.manager.testing
-        )
-
-    def authenticated_admin_from_request(self):
-        """Returns an authenticated admin or begins the Google OAuth flow"""
-
-        access_token = flask.session.get("admin_access_token")
-        if access_token:
-            admin = get_one(self._db, Admin, access_token=access_token)
-            if admin and self.google.active_credentials(admin):
-                return admin
-        return INVALID_ADMIN_CREDENTIALS
-
-    def authenticated_admin(self, admin_details):
-        """Creates or updates an admin with the given details"""
-
-        admin, ignore = get_one_or_create(
-            self._db, Admin, email=admin_details['email']
-        )
-        admin.update_credentials(
-            self._db, admin_details['access_token'], admin_details['credentials']
-        )
-        return admin
-
-    def signin(self):
-        """Redirects admin if they're signed in."""
-        admin = self.authenticated_admin_from_request()
-
-        if isinstance(admin, ProblemDetail):
-            redirect_url = flask.request.args.get("redirect")
-            return redirect(self.google.auth_uri(redirect_url), Response=Response)
-        elif admin:
-            return redirect(flask.request.args.get("redirect"), Response=Response)
-
-    def redirect_after_signin(self):
-        """Uses the Google OAuth client to determine admin details upon
-        callback. Barring error, redirects to the provided redirect url.."""
-
-        admin_details, redirect_url = self.google.callback(flask.request.args)
-        if isinstance(admin_details, ProblemDetail):
-            return self.error_response(admin_details)
-
-        if not self.staff_email(admin_details['email']):
-            return self.error_response(INVALID_ADMIN_CREDENTIALS)
-        else:
-            admin = self.authenticated_admin(admin_details)
-            flask.session["admin_access_token"] = admin_details.get("access_token")
-            flask.session["csrf_token"] = base64.b64encode(os.urandom(24))
-            return redirect(redirect_url, Response=Response)
-
-    def check_csrf_token(self):
-        """Verifies that the provided CSRF token is valid."""
-        token = self.get_csrf_token()
-        if not token or token != flask.request.form.get("csrf_token"):
-            return INVALID_CSRF_TOKEN
-        return token
-
-    def get_csrf_token(self):
-        """Returns the CSRF token for the current session."""
-        return flask.session.get("csrf_token")
-
-    def staff_email(self, email):
-        """Checks the domain of an email address against the admin-authorized
-        domain"""
-
-        staff_domain = Configuration.policy(
-            Configuration.ADMIN_AUTH_DOMAIN, required=True
-        )
-        domain = email[email.index('@')+1:]
-        return domain == staff_domain
-
-    def error_response(self, problem_detail):
-        """Returns a problem detail as an HTML response"""
-        html = self.ERROR_RESPONSE_TEMPLATE % dict(
-            status_code=problem_detail.status_code,
-            message=problem_detail.detail
-        )
-        return Response(html, problem_detail.status_code)
-
-
 class IndexController(CirculationManagerController):
     """Redirect the patron to the appropriate feed."""
 
@@ -545,8 +449,7 @@ class OPDSFeedController(CirculationManagerController):
 
     def groups(self, languages, lane_name):
         """Build or retrieve a grouped acquisition feed."""
-        admin = isinstance(self.manager.admin_controller.authenticated_admin_from_request(), Admin)
-        
+
         lane = self.load_lane(languages, lane_name)
         if isinstance(lane, ProblemDetail):
             return lane
@@ -556,13 +459,12 @@ class OPDSFeedController(CirculationManagerController):
 
         title = lane.display_name
 
-        annotator = self.manager.annotator(lane, admin=admin)
+        annotator = self.manager.annotator(lane)
         feed = AcquisitionFeed.groups(self._db, title, url, lane, annotator)
         return feed_response(feed.content)
 
     def feed(self, languages, lane_name):
         """Build or retrieve a paginated acquisition feed."""
-        admin = isinstance(self.manager.admin_controller.authenticated_admin_from_request(), Admin)
 
         lane = self.load_lane(languages, lane_name)
         if isinstance(lane, ProblemDetail):
@@ -573,7 +475,7 @@ class OPDSFeedController(CirculationManagerController):
 
         title = lane.display_name
 
-        annotator = self.manager.annotator(lane, admin=admin)
+        annotator = self.manager.annotator(lane)
         facets = load_facets_from_request()
         if isinstance(facets, ProblemDetail):
             return facets
@@ -588,7 +490,6 @@ class OPDSFeedController(CirculationManagerController):
         return feed_response(feed.content)
 
     def search(self, languages, lane_name):
-        admin = isinstance(self.manager.admin_controller.authenticated_admin_from_request(), Admin)
 
         lane = self.load_lane(languages, lane_name)
         if isinstance(lane, ProblemDetail):
@@ -603,7 +504,7 @@ class OPDSFeedController(CirculationManagerController):
 
         # Run a search.    
         this_url += "?q=" + urllib.quote(query.encode("utf8"))
-        annotator = self.manager.annotator(lane, admin=admin)
+        annotator = self.manager.annotator(lane)
         info = OpenSearchDocument.search_info(lane)
         opds_feed = AcquisitionFeed.search(
             _db=self._db, title=info['name'], 
@@ -896,13 +797,12 @@ class WorkController(CirculationManagerController):
         returns a single entry while the /works lookup protocol returns a
         feed containing any number of entries.
         """
-        admin = isinstance(self.manager.admin_controller.authenticated_admin_from_request(), Admin)
 
         pool = self.load_licensepool(data_source, identifier)
         if isinstance(pool, ProblemDetail):
             return pool
         work = pool.work
-        annotator = self.manager.annotator(None, admin=admin)
+        annotator = self.manager.annotator(None)
         return entry_response(
             AcquisitionFeed.single_entry(self._db, work, annotator)
         )
@@ -925,30 +825,6 @@ class WorkController(CirculationManagerController):
         data = flask.request.data
         controller = ComplaintController()
         return controller.register(pool, data)
-
-    def suppress(self, data_source, identifier):
-        """Suppress the license pool associated with a book."""
-        
-        # Turn source + identifier into a LicensePool
-        pool = self.load_licensepool(data_source, identifier)
-        if isinstance(pool, ProblemDetail):
-            # Something went wrong.
-            return pool
-    
-        pool.suppressed = True
-        return Response("", 200)
-
-    def unsuppress(self, data_source, identifier):
-        """Unsuppress the license pool associated with a book."""
-        
-        # Turn source + identifier into a LicensePool
-        pool = self.load_licensepool(data_source, identifier)
-        if isinstance(pool, ProblemDetail):
-            # Something went wrong.
-            return pool
-    
-        pool.suppressed = False
-        return Response("", 200)
 
 
 class ServiceStatusController(CirculationManagerController):
