+++ test
@@ -1,2 +1,2 @@
 #!/bin/sh
-nosetests -w tests
+nosetests tests
